+++ src/accounts/index.ts
@@ -377,7 +377,7 @@ function checkAddress(config: RhinestoneConfig) {
 }
 
 // Signs and packs a signature to be EIP-1271 compatible
-async function getPackedSignature(
+async function getEip1271Signature(
   config: RhinestoneConfig,
   signers: SignerSet | undefined,
   chain: Chain,
@@ -427,6 +427,26 @@ async function getPackedSignature(
       throw new Error(`Unsupported account type: ${(account as any).type}`)
     }
   }
+}
+
+// Signs and packs a signature to be used by the emissary validator
+async function getEmissarySignature(
+  config: RhinestoneConfig,
+  signers: SignerSet | undefined,
+  chain: Chain,
+  hash: Hex,
+  transformSignature: (signature: Hex) => Hex = (signature) => signature,
+): Promise<Hex> {
+  if (config.account?.type === 'eoa') {
+    throw new EoaSigningNotSupportedError('packed signatures')
+  }
+  signers = signers ?? convertOwnerSetToSignerSet(config.owners!)
+  const address = getAddress(config)
+
+  const signFn = (hash: Hex) =>
+    signMessage(signers, chain, address, hash, false)
+  const signature = await signFn(hash)
+  return transformSignature(signature)
 }
 
 // Signs and packs a signature to be EIP-1271 compatible
@@ -508,10 +528,6 @@ async function isDeployed(config: RhinestoneConfig, chain: Chain) {
   })
   if (!code) {
     return false
-  }
-  if (code.startsWith('0xef0100') && code.length === 48) {
-    // Defensive check to ensure there's no storage conflict; can be lifted in the future
-    throw new ExistingEip7702AccountsNotSupportedError()
   }
   return size(code) > 0
 }
@@ -913,7 +929,8 @@ export {
   toErc6492Signature,
   getSmartAccount,
   getGuardianSmartAccount,
-  getPackedSignature,
+  getEip1271Signature,
+  getEmissarySignature,
   getTypedDataPackedSignature,
   deployStandaloneWithEoa,
   // Errors